Add chat app menu button.
